additional classification models added
