+++ .gitignore
@@ -127,3 +127,6 @@ dmypy.json
 
 # Pyre type checker
 .pyre/
+
+# model files
+/models/classification_models/rf_ovr_classif.pickle
